+++ R/HIBP_headers.R
@@ -3,12 +3,17 @@
 #'
 #' @param agent The user agent the API call should run under
 #'
+#'
 #' @return headers
 #' @export
 #'
 #' @examples
 #' HIBP_headers()
-HIBP_headers <- function(agent="HIBPwned R pkg") {# nolint
+HIBP_headers <- function(agent = "HIBPwned R pkg") {# nolint
+  if (is.null(agent)){
+    agent <- "HIBPwned R pkg"
+  }
+
   if (length(agent) != 1 | !inherits(agent, "character")){
     stop("Problematic agent")
   }