+++ R/account_breaches.R
@@ -4,10 +4,14 @@
 #' @param truncate Whether responses should be trimmed to just name of
 #' breached site
 #' @param domain Search a specific domain
-#' @param as_list whether to return results as a list (default, TRUE) or a data.frame
-#' @param ... Optional passthrough to HIBP_headers()
+#' @param as_list whether to return results as a list (default, TRUE) or a data.frame,
+#'  When returned as a data.frame only accounts with at least one breach are returned.
+#' @inheritParams data_classes
 #'
 #' @return List of data.frames containing results
+#'
+#' @inherit data_classes details
+#'
 #' @export
 #'
 #' @examples
@@ -16,7 +20,9 @@ account_breaches <- function(
                              accounts
                              , truncate=FALSE
                              , domain=NULL
-                             , as_list = TRUE, ...) {
+                             , as_list = TRUE
+                             , verbose = TRUE
+                             , agent = NULL) {
 
   if (length(accounts) == 0 | !inherits(accounts, "character")){
     stop("Problematic accounts")
@@ -42,14 +48,17 @@ account_breaches <- function(
     URLS <- urltools::param_set(URLS, "domain", urltools::url_encode(domain))# nolint
   }
 
-  res <- lapply(URLS, GETcontent, HIBP_headers(...))# nolint
+  res <- lapply(URLS, GETcontent, HIBP_headers(agent), verbose = verbose)# nolint
   names(res) <- accounts
   if (!as_list){
     # bind the list and keep the account as a column
+    accounts <- accounts[unlist(lapply(res, ncol)) > 2]
+    res <- res[unlist(lapply(res, ncol)) > 2]
     account <- rep.int(accounts, times = lapply(res, nrow))
     if (length(res) > 1){
-      res <- do.call("rbind", res)
 
+      res <- do.call("rbind", res)
+      rownames(res) <- c()
     }else{
       res <- res[[1]]
     }